refactor(cli): restructure CLI as shared IPC client, address feedback
ADD:
- New :integration:client module with DesktopClient, DesktopLauncher, DesktopResult, PortResolver
- CliContext.kt with shared DesktopLauncher instance for all commands
- DefaultDesktopBinaryProvider with jpackage.app-path support and platform fallback
- Parameterized DesktopLauncher class with binaryProvider: () -> File? delegate

CHANGE:
- Move DesktopClient, DesktopResult from cli/app/client to :integration:client
- Rewrite DesktopLauncher from object to class with binary dependency injection
- Rewrite PortResolver to use shared MapConfig + intKeyOf infrastructure
- Update AddCommand/ListCommand/ControlCommands imports for CliContext
- Inline SizeConverter.bytesToSize() calls, remove CliFormatting formatSize/formatSpeed
- Replace System.exit() with exitProcess() in CliMain.kt
- Add :integration:client and :shared:config dependencies to build files

REMOVE:
- cli/app/client/DesktopLauncher.kt (replaced by DefaultDesktopBinaryProvider)
- cli/app/utils/PortResolver.kt (moved to :integration:client)
- CliFormatting formatSize/formatSpeed (delegated to SizeConverter)
- CLI.md Installation section
- abdm-cli.bat launcher script from installer resources
- AddToPath.nsh/RemoveFromPath.nsh from NSIS installer template
- CLI jar references from desktop build
