+++ cli/app/src/main/kotlin/com/abdownloadmanager/cli/commands/ControlCommands.kt
@@ -1,10 +1,10 @@
 package com.abdownloadmanager.cli.commands
 
-import com.abdownloadmanager.cli.client.DesktopLauncher
-import com.abdownloadmanager.cli.client.DesktopClient
-import com.abdownloadmanager.cli.client.DesktopResult
+import com.abdownloadmanager.cli.CliContext
 import com.abdownloadmanager.cli.utils.CliFormatting
-import com.abdownloadmanager.cli.utils.PortResolver
+import com.abdownloadmanager.integration.client.DesktopClient
+import com.abdownloadmanager.integration.client.DesktopResult
+import com.abdownloadmanager.integration.client.PortResolver
 import com.abdownloadmanager.integration.ApiDownloadModel
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.arguments.argument
@@ -13,6 +13,8 @@ import com.github.ajalt.clikt.parameters.options.*
 import com.github.ajalt.clikt.parameters.types.long
 import com.github.ajalt.mordant.rendering.TextColors
 import com.github.ajalt.mordant.terminal.Terminal
+import ir.amirab.util.datasize.CommonSizeConvertConfigs
+import ir.amirab.util.datasize.SizeConverter
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.builtins.ListSerializer
@@ -32,7 +34,7 @@ abstract class BaseControlCommand(
     final override fun run() {
         val term = Terminal()
 
-        if (!DesktopLauncher.ensureDesktopRunning()) {
+        if (!CliContext.desktopLauncher.ensureDesktopRunning()) {
             term.println((TextColors.red)("Error: AB Download Manager is not available."))
             return
         }
@@ -123,7 +125,7 @@ class InfoCommand : CliktCommand(
         val term = Terminal()
         val jsonParser = Json { ignoreUnknownKeys = true; prettyPrint = true }
 
-        if (!DesktopLauncher.ensureDesktopRunning()) {
+        if (!CliContext.desktopLauncher.ensureDesktopRunning()) {
             term.println((TextColors.red)("Error: AB Download Manager is not available."))
             return
         }
@@ -184,8 +186,8 @@ class InfoCommand : CliktCommand(
                     else -> TextColors.white
                 }
                 term.println((TextColors.brightBlue)("  Status:  ") + statusColor(item.status))
-                term.println((TextColors.brightBlue)("  Size:    ") + CliFormatting.formatSize(item.size))
-                term.println((TextColors.brightBlue)("  Downloaded: ") + CliFormatting.formatSize(item.downloaded))
+                term.println((TextColors.brightBlue)("  Size:    ") + SizeConverter.bytesToSize(item.size, CommonSizeConvertConfigs.BinaryBytes).toString())
+                term.println((TextColors.brightBlue)("  Downloaded: ") + SizeConverter.bytesToSize(item.downloaded, CommonSizeConvertConfigs.BinaryBytes).toString())
 
                 // Speed and progress are currently hardcoded to 0 in the API.
                 // ApiDownloadModel.kt defines speed and progress fields, but
@@ -196,7 +198,7 @@ class InfoCommand : CliktCommand(
                 // toApiModel() to read from ProcessingDownloadItemState.
 
                 // if (item.speed > 0) {
-                //     term.println((TextColors.brightBlue)("  Speed:   ") + CliFormatting.formatSpeed(item.speed))
+                //     term.println((TextColors.brightBlue)("  Speed:   ") + SizeConverter.bytesToSize(item.speed, CommonSizeConvertConfigs.BinaryBytes).toString() + "/s")
                 // }
                 // if (item.progress > 0) {
                 //     term.println((TextColors.brightBlue)("  Progress: ") + "%.1f%%".format(item.progress))
@@ -215,7 +217,7 @@ class InfoCommand : CliktCommand(
                     term.println((TextColors.brightBlue)("  Con——nections: ") + item.connections.toString())
                 }
                 if (item.speedLimit > 0) {
-                    term.println((TextColors.brightBlue)("  Speed limit: ") + CliFormatting.formatSpeed(item.speedLimit))
+                    term.println((TextColors.brightBlue)("  Speed limit: ") + SizeConverter.bytesToSize(item.speedLimit, CommonSizeConvertConfigs.BinaryBytes).toString() + "/s")
                 }
                 if (item.checksum != null) {
                     term.println((TextColors.brightBlue)("  Checksum: ") + item.checksum)
