--- desktop/app/resources/installer/abdm-cli.bat
@@ -1,2 +1,0 @@
-@echo off
-"%~dp0runtime\bin\java.exe" -jar "%~dp0cli\abdm-cli.jar" %*
