+++ settings.gradle.kts
@@ -24,6 +24,7 @@ include("desktop:mac_utils")
 include("downloader:core")
 include("downloader:monitor")
 include("integration:server")
+include("integration:client")
 include("shared:utils")
 include("shared:app")
 include("shared:compose-utils")
